Refactoring related to the synchronous underlying http library
The program was originally making an async data source by wrapping the
underlying http library with a future and then later a Task. This
resulted in a lot of pointless Tasks. In fact Fetch is perfectly capable
of running synchronous tasks using Query.sync instead of Query.async so
I refactored that and also made my own Scheduler with only 4 threads to
limit concurrency.